move classes with keywords to separate file
